+++ src/model/messages/pause-video-data.ts
@@ -1,0 +1,5 @@
+export interface PauseVideoData {
+    reason?: string,
+    time?: number,
+    isLegacyPlayer?: boolean
+}